Changed the logic of displaying Token info
Previously it was working on specific IDs which were parsed into links.

But it didn't work on GH Pages, so it was decided to place token information on one URL, and because of that I needed to find some way to transfer data (token ID) from MainPage to SpecificTokenPage.

After trying Redux and Context API, it was decided to use the easiest way possible: to save token ID in session storage.